Save the payment method from the payment verification callback
Under certain circumstances payment method was empty in the database, to prevent this get the verified payment method from the status query and save it.
While testing the previous a path to receive empty payment method into the database was discovered, changed to the javascript to not report empty paytment methods.