Implement delivery reporting for strictly virtual orders
Normal orders go through shipping phase and gets reported as shipped to Svea. Orders that have only virtual products (gift-cards) skip that step and those event hooks entirely. For this we add an event hook on order save and check if order is completed and used svea to pay and has no shipments and only virtual produts, then we take a special path to bypass a lot of validation of items shipped and tracking numbers to send a delivery report to Svea for electronically delivered order.