Creditmemo does not align with original OrderRows
When doing multiple Creditmemos on the same order the rows did not match the original orderrows if rows had been fully refunded previously. This was because the rows completely disappeared from the creditmemo and we got an offset in our count. Instead work from the original order to count the rows and check which ones are being refunded from the creditmemo.